finished find-intersection
this code is crap, lol.

TODO: Work on a better solution that is much shorter.
Idea: Just merge all input strings, iterate over each number and sort
them into a hashmap. Count the occurence of each number, if occurence of
the number is == number of strings from the beginning, we have found an
intersection.

Signed-off-by: Christian Rebischke <chris@nullday.de>